+++ agrirouter-api-java-api/src/main/java/com/dke/data/agrirouter/api/service/messaging/CloudOffboardingService.java
@@ -2,5 +2,4 @@ package com.dke.data.agrirouter.api.service.messaging;
 
 import com.dke.data.agrirouter.api.service.parameters.CloudOffboardingParameters;
 
-public interface CloudOffboardingService extends MessagingService<CloudOffboardingParameters> {
-}
+public interface CloudOffboardingService extends MessagingService<CloudOffboardingParameters> {}
