+++ agrirouter-api-java-api/src/main/java/com/dke/data/agrirouter/api/service/messaging/CloudOnboardingService.java
@@ -1,11 +1,5 @@
 package com.dke.data.agrirouter.api.service.messaging;
 
-import com.dke.data.agrirouter.api.dto.onboard.OnboardingResponse;
 import com.dke.data.agrirouter.api.service.parameters.CloudOnboardingParameters;
 
-import java.util.List;
-
-public interface CloudOnboardingService extends MessagingService<CloudOnboardingParameters> {
-
-  List<OnboardingResponse> onboard(CloudOnboardingParameters parameters);
-}
+public interface CloudOnboardingService extends MessagingService<CloudOnboardingParameters> {}
